Alerts refactor.
Changed the way that alerts are created.  There is now a show alerts function that fades the alert in and then sets a timer to fade the alert out.  I also changed the  visible function so that it fades the alert out which makes it less jerky when the alert is closed.